+++ ndk-blossom/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nostr-dev-kit/ndk-blossom
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.19
+
 ## 0.1.17
 
 ### Patch Changes